updated routes and controller
