+++ server/routes/scana.js
@@ -1,5 +1,5 @@
 const router = require("express").Router();
-const scanController = require("./controllers/scanController");
+const scanController = require("./controllers/scanDbController");
 
 // Matches with "/api/books" Activity 11
 router.route("/")
