+++ src/utils/API.js
@@ -2,19 +2,11 @@ import axios from "axios";
 
 export default {
   // Gets books from the Google API
-  getBooks: function(q) {
-    return axios.get("/api/google", { params: { q: "title:" + q } });
+  getScan: function(q) {
+    return axios.get("/api/scan", { params: { q: "title:" + q } });
   },
   // Gets all saved books
-  getSavedBooks: function() {
-    return axios.get("/api/books");
-  },
-  // Deletes the saved book with the given id
-  deleteBook: function(id) {
-    return axios.delete("/api/books/" + id);
-  },
-  // Saves an book to the database
-  saveBook: function(bookData) {
-    return axios.post("/api/books", bookData);
+  getSavedScan: function() {
+    return axios.get("/api/scandb");
   }
 };
